+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 - Added <podcast:liveItem> support ([#18](https://github.com/vnali/craft-studio/discussions/18)).
+- Added <podcast:socialInteract> support ([#19](https://github.com/vnali/craft-studio/discussions/19)).
 
 ## 0.13.0 - 2023-07-19